[C8] Split recipe.go by recipe section (#2226)
model/drawing/recipe.go was 682 LOC, over the 500-line ceiling. Split the
drawing-document recipe by section, leaving the top-level envelope +
orchestration in place; each file is now under the ceiling:

- recipe.go (126): the drawingRecipe envelope and build/marshal/apply/restore
  orchestration (buildRecipe, MarshalRecipe, ApplyRecipe, applyRecipeStruct,
  Sheets.restore, setActiveIndex).
- recipe_sheet.go (101): the sheet/format frame (sheetRecipe, sheetRecipeOf,
  restoreSheetContents, restoreBorder, restoreDims).
- recipe_view.go (161): views + crop fences.
- recipe_dimension.go (123): dimensions + point3 cell conversion.
- recipe_annotation.go (164): annotations + revision-table rows.
- recipe_sketch.go (77): drawing sketches + hatch regions.

Behaviour-preserving: every struct and function body is byte-identical, only
moved; imports narrowed per file (the detached restoreDims comment is kept
verbatim). Drawing corpus + archguard pass before and after with no golden
updates. Net delta of recognizers / tolerance constants / fallback sites /
type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2226.

(cherry picked from commit 653c0607cc82d60086bca9e49d46ff58e5c775fe)
